+++ src/main/java/com/drazisil/luckyworld/world/WorldHandler.java
@@ -4,17 +4,23 @@ import com.drazisil.luckyworld.BlockSave;
 import com.drazisil.luckyworld.BlockSaveRecord;
 import com.drazisil.luckyworld.LuckyWorld;
 import org.bukkit.*;
+import org.bukkit.block.Beacon;
 import org.bukkit.block.Block;
 import org.bukkit.entity.Player;
 import org.bukkit.event.block.BlockBreakEvent;
+import org.bukkit.generator.ChunkGenerator;
 import org.bukkit.potion.PotionEffectType;
 
-import static com.drazisil.luckyworld.BlockSaveRecord.CenterType.CENTER;
-import static org.bukkit.Bukkit.*;
+import java.util.logging.Logger;
+
+import static com.drazisil.luckyworld.BlockSaveRecord.CenterType.CENTER_OFFSET_Y;
+import static org.bukkit.Bukkit.createWorld;
+import static org.bukkit.Bukkit.dispatchCommand;
 
 public class WorldHandler {
 
     private LuckyWorld plugin = LuckyWorld.getInstance();
+    private Logger logger = plugin.getLogger();
 
     private Location newSpawnLocation;
 
@@ -24,10 +30,22 @@ public class WorldHandler {
     public WorldHandler() {
         // https://hub.spigotmc.org/javadocs/spigot/org/bukkit/WorldCreator.html
 
-        WorldCreator newWorldCreator
-                = new WorldCreator("new_world").copy(getWorld("world_nether"));
+        WorldCreator newWorldCreator = new WorldCreator("new_world");
 
-        //        newWorldCreator.type(WorldType.BUFFET);
+//        newWorldCreator.generator(new LWChunkGenerator());
+
+        // Dubug things
+        String generatorSettings = newWorldCreator.generatorSettings();
+        logger.info("generator settings: " + generatorSettings);
+
+        ChunkGenerator chunkGenerator = newWorldCreator.generator();
+        logger.info("chunk generator: " + String.valueOf(chunkGenerator));
+
+        WorldType worldType = newWorldCreator.type();
+        logger.info(String.valueOf("worldtype: " + worldType));
+
+        World.Environment worldEnv = newWorldCreator.environment();
+        logger.info("worldenv: " + String.valueOf(worldEnv));
 
         newWorld = createWorld(newWorldCreator);
 
@@ -51,7 +69,8 @@ public class WorldHandler {
                                       Player player) {
 
         // Fetch spawn location
-        Location spawnLoc = getSpawnLocation();
+        Location spawnLoc = getSpawnLocation().clone();
+        spawnLoc.setY(spawnLoc.getY() - 1);
 
         if (plugin.locationToString(location).equals(plugin.locationToString(spawnLoc))
                 && event.getBlock().getType() == Material.GREEN_STAINED_GLASS) {
@@ -64,28 +83,31 @@ public class WorldHandler {
     }
 
     public void generateSpawnPlatform() {
-        
-        // TODO: Make platform unbreakable
 
-        Block spawnBlock = newWorld.getBlockAt(getSpawnLocation());
+        // TODO: Make platform unbreakable
 
         BlockSaveRecord blocksToChange
                 = new BlockSaveRecord();
         blocksToChange.generateBlockSaveCube(newWorld, getSpawnLocation().clone(),
-                1, 5, 5, CENTER, 0, 0, 0);
+                1, 5, 5, CENTER_OFFSET_Y, 0, -1, 0);
 
         for (BlockSave blockSave: blocksToChange.getBlocks()) {
             blockSave.getBlock().setType(Material.QUARTZ_BLOCK);
         }
 
-        spawnBlock.setType(Material.GREEN_STAINED_GLASS);
+        Location returnBlockLocation = getSpawnLocation().clone();
+        returnBlockLocation.setY(returnBlockLocation.getY() - 1);
+        Block returnBlock = newWorld.getBlockAt(returnBlockLocation);
+        returnBlock.setType(Material.GREEN_STAINED_GLASS);
 
         // Create beacon
-        Location beaconBlockLocation = getSpawnLocation().clone();
+        Location beaconBlockLocation = returnBlockLocation.clone();
         beaconBlockLocation.setY(beaconBlockLocation.getY() - 1);
 
         Block beaconBlock = newWorld.getBlockAt(beaconBlockLocation);
         beaconBlock.setType(Material.BEACON);
+        Beacon beaconState = (Beacon) beaconBlock.getState();
+        beaconState.setPrimaryEffect(PotionEffectType.JUMP);
 
         // Open the sky
         for (int i = (int) (getSpawnLocation().getY() + 1); i < newWorld.getMaxHeight(); i++) {